refactor(reservations): extract lifecycle module
